==> 4_PostGIS_Local.ipynb <==
# Writing to local PostGIS db

import pandas as pd
import os

db_host = os.environ.get('DB_HOST')
db_port = os.environ.get('DB_PORT')
db_user = os.environ.get('DB_USER')
db_password = os.environ.get('DB_PASSWORD')
db_name = os.environ.get('DB_NAME')

db_url = f'postgresql://{db_user}:{db_password}@{db_host}:{db_port}/{db_name}'

# read the csv file into a pandas dataframe, replacing NaN with None
df = pd.read_csv('final_output.csv', index_col=['cep_id'], dtype={'pa': 'Int64', 'eco': 'Int64', 'country': 'Int64'}).replace({pd.NA: None})
df.head()

#create table from dataframe 
from sqlalchemy import create_engine
engine = create_engine(db_url)
df.to_sql('cep_water', engine, if_exists='replace')

import psycopg2 as pg
from psycopg2 import sql
import pandas as pd


# connect to the database
conn = pg.connect(
    database=db_name,
    user=db_user,
    password=db_password,
    host=db_host,
    port=db_port
)


# create a cursor object using the cursor() method
cursor = conn.cursor()

# execute a statement
cursor.execute('SELECT * FROM cep_water LIMIT 5')

# print the result and close the connection
result = cursor.fetchall()
for row in result:
    print(row)

# close the cursor 
cursor.close()

def get_summed_bands_by_col(conn, group_by_col = "country_name",is_pa=False, is_marine=False, convert_to_meters=False, ):
    try:
        # create a cursor object using the cursor() method
        cursor = conn.cursor()
        # query all rows transition_0  to transition_10, filter where pa is 0 and where marine is false, and group by country_name and sum the areas for each transition band
        bands_SUM = [f'SUM(transition_{i}) as "transition_{i}"' for i in range(11)]
        query = sql.SQL('SELECT {group_by_column}, {bands} FROM cep_water WHERE is_protected = {is_pa} AND is_marine = {is_marine} GROUP BY {group_by_column}').format(
            bands=sql.SQL(', ').join(map(sql.SQL, bands_SUM)),
            is_pa=sql.Literal(is_pa),
            is_marine=sql.Literal(is_marine),
            group_by_column=sql.Identifier(group_by_col)
        )
        # #print query as string
        # print(query.as_string(conn))                                                                                                                    
        # filter where 
        cursor.execute(query)
        result = cursor.fetchall()
        cursor.close()

        # convert to df and set index to group_by_col
        df = pd.DataFrame(result, columns=[group_by_col]+[f'transition_{i}' for i in range(11)])
        df.set_index(group_by_col, inplace=True)
        # convert to meters
        if convert_to_meters:
            df = df.apply(lambda x: x/1000000)
        return df
        
        #return result
    except Exception as e:
        print("error: ", e)
    finally:
        conn.rollback()

df_protected = get_summed_bands_by_col(conn, group_by_col = "country_name",is_pa=True, is_marine=False, convert_to_meters=True)
df_unprotected = get_summed_bands_by_col(conn, group_by_col = "country_name",is_pa=False, is_marine=False, convert_to_meters=True)
df_terrestrial_eco = get_summed_bands_by_col(conn, group_by_col = "eco",is_pa=False, is_marine=False, convert_to_meters=True)

df_protected

df_unprotected

# close the connection
conn.close()
