==> 5_PostGIS_Cloud.ipynb <==
# Writing to Clever Cloud PostGIS db

import psycopg2 as pg
from psycopg2 import sql
import pandas as pd
import os

db_host = os.environ.get('DB_HOST_CLOUD')
db_port = os.environ.get('DB_PORT_CLOUD')
db_user = os.environ.get('DB_USER_CLOUD')
db_password = os.environ.get('DB_PASSWORD_CLOUD')
db_name = os.environ.get('DB_NAME_CLOUD')

db_url = f'postgresql://{db_user}:{db_password}@{db_host}:{db_port}/{db_name}'

# #create table from dataframe - Clever Cloud
# import pandas as pd
# # read the csv file into a pandas dataframe, replacing NaN with None
# df = pd.read_csv('final_output.csv', index_col=['cep_id'], dtype={'pa': 'Int64', 'eco': 'Int64', 'country': 'Int64'}).replace({pd.NA: None})
# from sqlalchemy import create_engine
# engine = create_engine(db_url)
# df.to_sql('cep_water', engine, if_exists='replace')

# connect to the database
conn = pg.connect(
    database=db_name,
    user=db_user,
    password=db_password,
    host=db_host,
    port=db_port
)

# check size of database
cursor = conn.cursor()
cursor.execute('SELECT * FROM cep_water LIMIT 5')
result = cursor.fetchall()
for row in result:
    print(row)
cursor.close()

conn.close()
